moved OpenAPI generator to correct namespace
added missing doc blocks
